+++ src/Http/Controllers/Dominion/AdvisorsController.php
@@ -235,6 +235,7 @@ class AdvisorsController extends AbstractDominionController
         return view('pages.dominion.advisors.statistics', [
             'landCalculator' => app(LandCalculator::class),
             'militaryCalculator' => app(MilitaryCalculator::class),
+            'opsCalculator' => app(OpsCalculator::class),
             'populationCalculator' => app(PopulationCalculator::class),
             'targetDominion' => $target
         ]);